requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
Django==6.0.4
django-apscheduler==0.7.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
